implemented slugify for content urls and added minor updates
