+++ pages/allotjaments.js
@@ -34,6 +34,7 @@ const PlaceList = ({ user }) => {
       <PublicContentBox
         key={el._id}
         type={el.type}
+        slug={el.slug}
         id={el._id}
         image={el.cover}
         title={el.title}
