+++ styles/globals.scss
@@ -1254,7 +1254,7 @@ nav {
     max-height: 40px;
     border-radius: 8px;
     padding: 0;
-    margin: 0 43px;
+    margin: 12px 43px;
   }
   .image img {
     width: auto;
